requirements.txt
Django==5.2.9
django-allauth==65.13.1
PyJWT==2.10.1
cryptography==46.0.3
Pillow==12.0.0
PyMuPDF==1.26.7
python-dotenv==1.2.1
google-generativeai==0.8.6
requests==2.32.5
requests-oauthlib==2.0.0
gunicorn==23.0.0
whitenoise==6.11.0
python-decouple==3.8
dj-database-url==3.1.0
psycopg2-binary==2.9.11
oauthlib==3.3.1
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
urllib3==2.6.2
sqlparse==0.5.5
asgiref==3.11.0
cffi==2.0.0
pycparser==2.23
cloudinary==1.41.0
django-cloudinary-storage==0.3.0
